feat: move server-only files to lib/server
Svelte-kit protects these files so that they cannot be read by the client. Also, we create a config folder for instance-specific configs.
