+++ frontend/tests/getData.test.ts
@@ -2,7 +2,7 @@
 // @ts-ignore
 import {describe, test, expect, vi, Mock} from 'vitest';
 import {getAllCandidates, getData} from '$lib/api/getData';
-import {constants} from '$lib/utils/constants';
+import {constants} from '$lib/server/config/constants';
 import candidates from './data/candidates.json';
 import parties from './data/parties.json';
 import singleCandidate from './data/singleCandidate.json';
